declutter save.h (items, spellbooks)
